Move ConfigSubstitutionTests to shared
